Fix errors "'./benchmarks/benchmarks/lib/*.jar': No such file or directory"
If 'benchmarks/benchmarks/lib/*.jar' don't exist the loop processing jar
files gets an invalid input list. This causes the jar tool to fail.

This CL fixes the bug by checking if jar files exist.

Test: benchmarks/build.sh -t

+++ build.sh
@@ -194,13 +194,18 @@ JAVA_FRAMEWORK_FILES="$(find $DIR_FRAMEWORK -type f -name '*'.java)"
 verbose_safe rm -rf $DIR_BUILD
 verbose_safe mkdir -p $DIR_BUILD/classes/
 
-for jar_file in "${DIR_BENCHMARKS}"/lib/*.jar
-do
-  jar_file="$(realpath "${jar_file}")"
-  # Extract jar file and remove META-INF, which is not needed and can cause
-  # issues with target runs.
-  (cd $DIR_BUILD/classes && jar xfv "${jar_file}" && rm -rf META-INF)
-done
+jar_files=("${DIR_BENCHMARKS}"/lib/*.jar)
+# The ${DIR_BENCHMARKS}/lib directory might contain jar files needed for benchmarks.
+# We need to check whether there are any jar files before trying to process them.
+if [[ -f "${jar_files[0]}" ]]; then
+  for jar_file in "${jar_files[@]}"
+  do
+    jar_file="$(realpath "${jar_file}")"
+    # Extract jar file and remove META-INF, which is not needed and can cause
+    # issues with target runs.
+    (cd $DIR_BUILD/classes && jar xfv "${jar_file}" && rm -rf META-INF)
+  done
+fi
 
 javac_cmd_options=("-encoding" "UTF-8" \
   "-cp" "${DIR_BENCHMARKS}:${DIR_BUILD}/classes" \
